requirements.txt
phi
langchain
groq
python-dotenv
uvicorn
fastapi
duckduckgo-search
yfinance
